optimization

+++ src/ray-tracer-sdl2-win32/tests.cpp
@@ -257,43 +257,42 @@ void test_matrix_cofactor()
 void test_matrix_inverse()
 {
 	// Set up
-	Matrix<float, 4, 4> m0 {
+	Matrix<double, 4, 4> m0 {
 		-8,-5, 9, 2,
 		 7, 5, 6, 1,
 		-6, 0, 9, 6,
 		-3, 0,-9,-4
 	};
 
-	Matrix<float, 4, 4> m0_inverted{
+	Matrix<double, 4, 4> m0_inverted{
 		-0.15385, -0.15385, -0.28205, -0.53846,
 		-0.07692,  0.12308,  0.02564,  0.03077,
 		 0.35897,  0.35897,  0.43590,  0.92308,
 		 0.69231, -0.69231, -0.76923, -1.92308
 	};
 
-	Matrix<float, 4, 4> m1_error{
+	Matrix<double, 4, 4> m1_error{
 		 -4, 2,-2,-3,
 		  9, 6, 2, 6,
 		  0,-5, 1,-5,
 		  0, 0, 0, 0
 	};
 
-	Matrix<float, 4, 4> a {
+	Matrix<double, 4, 4> a {
 		 3, -9,  7,  3,
 		 3, -8,  2, -9,
 		-4,  4,  4,  1,
 		-6,  5, -1,  1
 	};
 
-
-	Matrix<float, 4, 4> b {
+	Matrix<double, 4, 4> b {
 		 8,  2, 2, 2,
 		 3, -1, 7, 0,
 		 7,  0, 5, 4,
 		 6, -2, 0, 5
 	};
 
-	Matrix<float, 4, 4> result, c, inverse_b, mul_a;
+	Matrix<double, 4, 4> result, c, inverse_b, mul_a;
 
 	// Assert
 	assert(!matrix_inverse(m1_error, result));
